Whew.  Got a bit out of sync there.  Tons of stuff in this update.
Minor highlights are:

  * Unit tests pass under 1.5.2 again.

  * tkinternet works again.

  * Minor GUI updates for faucets & IM

And of course, the real reason for this update, and all the debugging I've been
doing:

  * Restructuring of PB to decouple authentication from the protocol.

This breaks the protocol (obviously).  However, this means you can publish any
sort of referenceable object on a port, not just authenticated ones.  This
makes the web distribution stuff a lot simpler.  (And also renders it
backwards-incompatible.  Yay. :-\)

This also comes with a few utility methods which automate the tedious
state-machine which you need in order to log in properly.  I deleted a bunch of
ad-hoc implementations of this and replaced them all with a common one (which
is still a little too ad-hoc for my tastes, but at least it's in a central
location now).  Unsurprisingly this is an overall reduction of code; it should
also make scripting of interactions with PB services a lot easier (in fact, one
of the reasons I did it is I wanted to cite a simple scripted interaction
example for my PB paper, and I couldn't write one in few enough lines of code).
Writing new PB clients will be easier, too.

Some of the names here are still sort of temporary.  I don't like 'AuthRoot'
and 'AuthServ', but they're all I can think of; I've named far too many classes
lately.

So without further ado, here it is.