First pass at asynchronous perspective access.
There are a few things not yet fixed, most notably twisted.reality and
twisted.words.  While both are updated to work with the new API (PB requires
it) I'm not quite sure how to fix 'em further; as they don't load their
perspectives from a database, there is a fair amount of code that assumes they
are synchronously accessible.

Twisted.web.guard also got an overhaul, and I integrated the patch that fixes
widgets' ordering of stuff (although there is a block commented out that is
supposed to reduce the length of lists by concatenating strings).

Twisted.Forums also authenticates out of a database now, and provides an
example of Twisted.Web.Guard.

Fixed some of itamar's changes to allow variatic args in the database
connector.  runQuery still needs that sort of flexibility (thankfully it's not
applicable to runInteraction) but it wasn't immediately obvious how to fix it
without making it super-inconvenient.