shutdown(2) sockets instead of close(), fixing bug reported by Drew.
The problem with close() is that it doesn't actually close the socket in
all cases. If we start a win32 process, it will inherit the socket handle
and now close() will only decref the socket in the TCP/IP stack. When
the subprocess closes the refcount will hit 0 and the socket will be
reset, ungracefully. If we do shutdown(2) however the socket is closed
regardless of its refcount, as intended.