Better collection support for coil.
1) Collections of immutable items (e.g. username/password dictionary) can
now be configured via COIL. See t.c.plugins.ftp for an example.

2) If a configurable class is not collection, coil's web interface will
check if the Configurator class is a collection, allowing us to have
Configurators that act as collections.

A remaining issue is twisted.web.vhost, and other similar cases, where
both the Configurator and the configurable class should really be
collections.