And now, the sequel (to the sequel) you've all been waiting for:
                    Bride of Reactor Refactor!!

TCP CALLS abstracted into a SINGLE OBJECT!  LOVE in a WORLD GONE MAD!
GLIB_MAIN sitting next to SELECT(2)!  CODING without UNIT TESTS!  This checkin
HAS IT ALL!!

For those of you just now joining the exciting Reactor Refactor storyline,
here's what's going on:

Pretty much all of the twisted.internet.* classes are being deprecated in favor
of a single, central class called the "reactor".  Interfaces for what this does
are documented in twisted.internet.interfaces; the IReactor* interfaces are the
interesting bits.

This checkin gets us about 3/4 of the way there.  All of the main-loop plugins
should work with the new scheme now (and not all of them are necessarily
reactors).  It is now possible to select one at startup with twistd --reactor.

Why is it called the "Reactor Refactor"?  This pattern is called "Reactor" in
POSA, and elsewhere.  We are refactoring Twisted to use it.

Why is this happening?  Becuase the mechanism used to plug-in different Twisted
main-loops was ... well, mostly it wasn't anything.  There was no way to change
loops, and no way to query the current main-loop for capabilities (such as SSL
support) was catching ImportErrors.  Also, it was impossible to instrument the
mainloop to get a view on what a Twisted server was currently doing.

What does this mean for you?  Well, if you are writing Twisted servers and
obeying the whole mktap/twistd paradigm, probably not much.  If you're writing
your own startup script, start adhering to that paradigm and things will work
again :-).  If you're writing a client in GTK, you will have to change your
initialization to install the GTK reactor and run it, rather than calling
gtk.mainloop/gtk.mainquit.  (See t-im for an example.)  If you are writing a tk
client, you'll have to install some reactor, then add the tk mainloop to it --
this pretty much just adds one line to your application (see im for an
example).

Also, this will allow us to much more easily maintain the documentation for
twisted.internet, since there are a lot of implementations of the same,
previously undeclared, interfaces lying around.

Jython now runs twistd much more reliably.  (Manhole is an especially useful
application of this, for those of you writing Java server-side applications!)

Right now, most code written to the old interfaces will work, but that will
change, so start reading the interfaces documentation right now.