Marmalade: Jelly, with just a hint of bitterness.
This is an XML object serialization module, designed to give system
administrators (and others) a "human" editable configuration file for Twisted.
Soon there should be some way for an object to control what sort of XML it
generates, so as to reduce the verbosity and complexity of the configuration,
as well as support for twistd to spit out .TAX (Twisted Application XML) rather
than .TAP (Twisted Application Pickle) files.