WHEN WILL THIS HAPPY-GO-LUCKY MERRY-GO-ROUND OF EXCITEMENT EVER END!?
Added --xml option to twistd + mktap, which allows them to read/write files in
marmalade format.  This obviously needs to be more thorough but it gives you a
basic idea of how it can be used.  Fixed a bug in jelly and in marmalade that
was causing instance methods to get screwed if their im_self was a
backreference.  Refactored circular reference utilities to be common between
jelly and marmalade.  Added unit tests to make sure that the method bug doesn't
crop up again.  Fixed a buglet in row.py that prevented it from being used in
marmalade.  Discovered that PyPgSQL invents its own whole typesystem that is
two star-systems away from Python's built-in type system.  What the hell?
Something's probably still wrong with marmalade anyway, since pickle can handle
pickling a reflector, but it can't. (blows up on a 'PyPgInt2: None')