Chris Armstrong's "PB Failures" patch
This is the best change to PB ever.  Added a fairly major feature, at the
protocol level, and didn't break backwards compat.!  (Well, at least it isn't
supposed to.  There are probably some _highly obscure_ theoretical cases where
it does break, but then, failure support was so lame before that it's unlikely
anybody was using it.)

This essentially gives a defined way to check for failure in a PB method call
that's quite a bit more robust than string matching ;-).  It also makes PB
depend sort of implicitly on the Python standard error types, so that will
probably have to be in the protocol documentation.