Updated and added tests for twisted.sister.  It may actually be usable for something now.
Note: With this checkin we (sean and I) discovered that the unit tests are
starting to interfere with each other.  Currently a full run of the tests will
not complete, and will in fact hang and be unresponsive to signals.  Running
individual test cases by themselves work fine, but they're interfering with
each other by doing reactor-y things ... we need each test case to
instantiate/start/stop the reactor, etc.  I will buy chocolate for whoever does
this first.