Initial add for the cReactor.  Changed twistd and setup.py to use and build the cReactor.
Moved IFactory out of twisted.internet.protocols and into
twisted.internet.interfaces and renamed it to IProtocolFactory.  Changed all
usages of IFactory to IProtocolFactory.

Added a few __implements__ class attributes and a few "if implements()" checks
instead of assuming a reactor will implement a given method.

Removed the server parameter from Protocol.makeConnection as it needed to go
(glyph said so).