Make session cookies not be path-specific, but global to the whole host. It should probably be "path that is the root of the namespace this instance controls", but this is better than with totally path-specific cookies.
Additionally, some confusion in Galeon or Twisted lead not just to
reauthenticating when changing the URL, but reauthenticating _every_
_single_ _time_.

Used to be: if two URL paths have ResourceGuards, you need to
reauthenticate on every page, e.g. submitting forms to ResourceGuarded
Forms was impossible due to this.

Is now: you need to reauthenticate on every page _if and only if_ the
one host name is really two different servers, both using twisted
sessions.