Acceptance tests: they're not just a good idea, they're the *law*.
I fixed two bugs here that were obviously caused by people not checking to see
if their changes to core code were correct with the test suites we've got:

  * Default SSL context didn't work with released versions of mozilla due to a
    version problem

  * UNIX sockets were totally fucked

Let's try to at least get the code to the point where we're not getting
UnboundLocalErrors before we check in, hmm?  (And yes, you're exempt from my
wrath right now if you're on windows.  But still, run pychecker or something!)