Change LogFile to support disabling log rotation if rotateLength is None.
We probably want to add an option to twistd, say --logsize, to be able to
control this from the command line.

The apt-proxy guys have a problem where trying to rotate logs in a directory
where you don't have permissions (say we dropped privileges after starting)
causes twisted to silently bomb out :(... this should help them work around
that.  Find haggai or ranty on IRC for more info.