Added API to domwidgets.Widget; generate(self, request, node) and setUp(self, request, node, data)
generate can be implemented to return any of the types handled by DOMTemplate.dispatchNode; a string, a node, a web.widgets.Widget, a DOMWidget, or a Deferred

However, more interesting for the implementation of complex widgets, setUp can be overridden (replacing the behavior of overriding generateDOM to do setup) to do anything required to set up the widget before rendering. As an added bonus, the Model data the Widget is meant to operate on is passed in the fourth parameter to setUp, data.

Todo: Refactor getData and setData to be more safe; make it easier to create Widgets which operate on heirarchies of data by refactoring List's handling of model attributes into a more general solution; refactor all the Widgets in domwidgets to use setUp instead of overriding generateDOM.