No more eval or exec in domwidgets; horray!
Added getSubmodel and setSubmodel api on mvc.Model; changed getData and setData to use these APIs; wrote a wrapper for simple Lists that allows someone to say <div model="foo/0" view="Text" /> and get the first element of the list named "foo" on the model. Registerred the wrapper as implementing IModel for types.ListType.