(IRCClient): Document instance variables.
(IRCClient.ctcpQuery_SOURCE): Decided that the document which
assumes the client's source code is available by anonymous FTP is
just *wrong*, and to screw their funky "host:dirname:file" syntax
and send a URL instead.

(IRCClient._sendLine): I monkied with the delayed send queue.
Made it a little cleaner, to my mind anyhow, without the dance of
the self.sendDelayed attribute.  Also prefixed the queue attribute
with a _ to discourage folks from messing with it.

Marking myself as maintainer.  Is this what my life has come to?