Slight changes to make getPattern() infinitely recurse less often.
This change is required for certain kinds of nested listing to work properly.
In particular, this example:

    <a view="List">
        <b pattern="listItem" view="List>
            <c pattern="listItem" /> </b> </a>

blows up horribly the way that it used to work.  I was a little confused to see
that _get uses id, model, and pattern, and I don't think that should be the
case any more, but I generalized that a little bit.

Currently what it's doing is slightly better (it won't return the C node in A's
getPattern, nor the B node in B's getPattern) but I still have some concerns.
With many models (all list models, as far as I know) being anonymous, it's not
enough to say listItemOf="modelname".  A more appropriate solution might be to
require an id="" attribute for any nodes that use patterns, but then we run
into the problem of determining whether the user assigned the id attribue or
woven did before determining whether it's more appropriate to use the 'model'
or 'id' attribute in order to locate the pattern.  I didn't know how to do that
so I haven't implemented it yet.

I don't know if it's a good idea to *ever* allow patterns without some unique
identifier for each level... consider this situation:

    <x view="List">
        <y pattern="listHeader" />
        <z pattern="listItem" view="List" />
            <q pattern="listItem" />
            <r pattern="listFooter" />
        </z>
    </x>

Is there *any* way for woven to tell that R belongs to Z, and not to X?

More to follow, I'm sure.