Only count the first failure for any given test.
Any failures in the test case, tearDown, or the reactor-cleaning afterwards
will flunk the test. But only put the first one into the output log,
otherwise multiple failures can report more errors than we have tests. The
second or third failure is generally an artifact of the first one anyway.