Newer, shinier overview diagram!
The primary motivation for this was I needed a .dia source file for decent
generation of the .ps format Twisted docs.  Plus, the old diagram was out of
date (Twisted Reality isn't a core feature any more), and had the colour palette
of a Commodore 64... ;)

This new image looks much sexier in the .ps docs, but the .png for the web is a
little bit lacking.  I'm going to look at using gs to render the .png from the
.eps, which should make it much nicer.