breaking t.im; start small.
WARNING: The next person to work with twisted.im, be it radix, myself,
or someone else, is most likely going to break shit, because this
package needs some reconstructive surgery.  Currently, there are NO
TESTS for this package, nothing to tell us how much we've broken.  And
if we don't know how much we've broken, there's no reason to feel
guilty about it, right?

The upshot of this is that if you're using this package and there's
something that you need not to break, or maybe broken a little but at
least not be rewritten out of existance, WRITE TESTS NOW.

There are two API changes in this commit:
 1) getStatus methods return the status constant from im.locals, not a string.
    If you need a string, str() the result.

 2) Account.startLogOn is now Account._startLogOn, as that's an implementation
    detail of the base's implementation of logOn and not part of Account's
    public interface.


Now back to our regular program:

* twisted/im/interfaces.py: Attempt to define interfaces for this
package.

* twisted/im/locals.py: use more unique and examinable objects for
status constants than anonymous 0,1,2 values.

* twisted/im/basesupport.py (AbstractAccount._startLogOn): renamed
from .startLogOn to ._startLogOn;
also, epydocify.

* twisted/im/pbsupport.py,
* twisted/im/ircsupport.py: changes for status enum and _startLogOn

* twisted/im/gtkchat.py (ContactsList.refreshContactsLists): use str() of status.