Deferredization of im logOns.
I'm not confident that the client's constructor is the best place to
pass in the deferred, but it seemed the most workable of the
alternatives I came up with.

I succeeded in small part in using deferreds to cut down on the
excessive intertwining of accounts, clients, and the ChatUI, in that I
got rid of Account.registerAs, but there's still a long way to go.


* twisted/im/basesupport.py (AbstractClientMixin.__init__): 1.13 add
logonDeferred
(AbstractClientMixin.registerAsAccountClient): 1.13 gone, moved to
(AbstractAccount.logOn): 1.13 register and return the deferred
client.  I'm not especially keen about registering it here either --
IMHO if the caller is the GUI, it can take care of that itself
-- but this is the path of least breakage.

* twisted/im/interfaces.py (IClient.__init__): 1.3 Accept a
logonDeferred parameter.  The client should call this back when
it is successfully logged in.

* twisted/im/basechat.py (ChatUI.registerAccountClient): 1.6
* twisted/im/baseaccount.py (AccountManager.connect): 1.3 return
the deferred client.

* twisted/im/tocsupport.py (TOCProto.__init__): 1.21
(TOCProto.tocNICK): 1.21
(TOCAccount._startLogOn): 1.21
* twisted/im/ircsupport.py (IRCProto.__init__): 1.25
(IRCProto.connectionMade): 1.25
(IRCAccount._startLogOn): 1.25
* twisted/im/pbsupport.py (TwistedWordsClient.__init__): 1.23
(TwistedWordsClient.connected): 1.23 do logonDeferred happy

(PBAccount.logOn): 1.23 logonDeferred happy gets a little more
complicated here, as apparently there's a one-to-many relationship
between Accounts and Clients

* twisted/internet/protocol.py (ReconnectingClientFactory):
speeling fix, s/inital/initial/