FileDescriptors are Ephemeral
I was hunting a pickling error.  Something was holding a reference to
list.remove, which pickle didn't like.  That turned out to be the
canceller on a DelayedCall.  The DelayedCall was referenced by the
Reactor.  The Reactor was referenced by a Port, by code in its File
Descriptor parent.


<Acapnotic> dunno, tcp.Port has a __getstate__, documented as "(internal) get
my state for persistence"

<itamar> kill it