cleanup FILTH, more useful progress and debug params, fix debug behavior
But I still see no way to make --progress work.  It's probably not too
important, but it would be nice.

Actually I just thought of a way.  Double-pass parsing, with two Options
subclasses.  Brilliant.  I'll check it in tomorrow.

Also added w/ this commit is a generalized progress bar function and some
deprecations in plugin.py