make failures much faster in common case
what we do is, if an exception object was explicitly passed in, we
don't build a traceback. you want a traceback, pass that in too, or do:

try:
   foo()
except FooError:
   f = Failure()

and you'll get one. But this way e.g. we don't get massive stack
introspection for each lost connection. Anyone feel like cleaning up
the stack/frames code? it's hideous.