Put less strenuous ("insane") demands on the Mailbox
Changed the API.  listMessages() now takes an optional argument that is the
index of the message about which to return information.  This allows
commands that operate on a single message to only require *that* single
message to be loaded, rather than loading *all* messages!

I suspect I will want to change this API even further.