stage.result => stage.results
Implemented exarkun's suggestion.  This is quite a massive change in logic
as it moves towards dealing with multiple rows at the same time.  This is
quite a bit more complicated protocol... so I'm sure there are a few bugs.
In particular, the function calling order is critical to maintain, this
will make it perhaps a bit harder to write stages.  However, overall, it
is quite an improvement both in speed and in overall logic... I think.