[I]UsernamePassword that can support APOP (at least) and Deferred bug fix
Dang, dogg, the unit tests take a loooong time to run.

IUsernamePassword needs to specify how to talk to a username/password database
that knows how to store usernames and plain-text passwords.  This should
basically be able to handle any kind of challenge/response authentication.
UsernamePassword should probably implement more interfaces than
IUsernamePassword, and the in-memory example should cope with those interfaces
as well, but we will have to wait to implement other credential schemes to see
what those will shape up to look like.

The important thing here is that a CredentialChecker should be based primarily
about what kind of database it's going against.

The Deferred fix here changes the semantics in a horrible way which I wish it
did not, however, the old semantics were wrong.  Previously, Deferred would set
isError to True only if an exception was raised OR errback was called.  It
would then set isError back to False UNLESS the returned value of each
subsequent callback was a Failure.  Also, gatherResults was _horribly_ broken
and had semantics which could not possibly make sense.

The only "real" test that this made fail was an FTP test, which SPIV SHOULD FIX
ALONG WITH THE REST OF FTP.