Anders Hammarquist's wonderful parsing changes
  This patch changes the way input is parsed on the server side.
<state>_<COMMAND> class attributes contain tuples beginning with the handler
method for <COMMAND>; subsequent elements are either callables which parse
part or all of a received line OR non-callables.  When the last callable in
the tuple has finished consuming the remainder of the line, or the line is
completely consumed, the handler is called with the return value of all the
callables and all the non-callables.

  Parsers return two-tuples or Deferreds.  The first element of the
two-tuple is the object passed to the handler; the second element is the
portion of the string unconsumed by the parser.

  As a result, messages (via APPEND -- the only command through which a
client can send a message to a server) need not be buffered in memory.  When
the client code is changed to a similar pattern, the buffering problems
there should disappear as well (though the solution may be more complex on
the client side, as messages may appear embedded within the list expression
syntax IMAP4 uses).

  Also changed are how MessageSets work, how the modified UTF-7 codec is
implemented, the IMessage interface (grew a method) and empty strings are now
correctly identified as requiring quoting.  Other minor changes present as
well.