Form Yumminess!
Updated woven form to merge models with templates.  For example, if the model
has default values but the template has nice formatting and positioning,
replaces the template DOM node with a happy merged DOM node containing the
same formatting information as well as default values.

Also made formatting of radiogroups and checkboxes nicer - number of rows
and columns can be specified so that long lists of checkboxes/radiogroups can
be formatted into tables.

Fixed problem in formmethod so that radiogroups will no longer die if defaults
are not specified.

Added bit of documentation to formmethod and form.  A very little bit.

Transcended my mortal limitations to  BECOME  A  GOD