The best thing to happen to woven since sliced bread
Thanks to Spiv's test, which attempted to render more than one view
based on the same deferred model, I came to the conclusion that
Woven's Deferred handling was COMPLETELY WRONG AND UTTERLY BROKEN.
It worked when there was only one view rendering a Deferred model
because there were no additional callbacks being added and thus
it wasn't a problem that the final callback in the chain was returning
None.

Then, I took a moment to revisit the design decision that I made
for DOMTemplate which allowed the application programmer to return
any type of value from a generate method and insert it into the DOM
using an INodeMutator adapter. Since Modern Woven programmers don't
need this, I realized how I could restructure the code slightly to
simply check for Deferred results or node results. The code thus gets
simpler and easier to understand, because NODE MUTATORS ARE COMPLETELY
UNNECCESSARY.

I also took this opportunity to apply Rich Cavanaugh's patch, which
provides an IView adapter for DeferredWrapper, so you do not have to
explicitly state the name of the view you would like to eventually render
the deferred.

THANK YOU ANDREW AND RICH

Closes issues 91 and 126.