the tests still pass, and the man said check it in
As it transpired on IRC:

    <glyph> fzZzy: should I check in that '.'-model hack for List?
    <glyph> the <span pattern="listItem" model="." /> hack, I should say
    <fzZzy> glyph: hmm. I think so.

This is a slight workaround for the default behavior for the 'listItem' pattern
ndoe with respect to View lookup.  By default (for reasons that we will not
discuss in depth here), when a List encounters a list item, it will assign a
view of 'DefaultWidget' to that node, effectively cancelling any display logic.
This can be handy if you are displaying a table and want to do something like

    <tr pattern="listItem"> <td model="." /> </tr>

which means "the <tr> is what's repeated but the td is where you put the data"

This allows you to say

    <div pattern="listItem" model="." />

and get the same default-view-adapter-lookup on the actual pattern node as you
could get on one of its subnodes.

I hope that makes sense.