Change NotImplementedError to -1 return.  Resolve 525.
When a FileDescriptor is connected to something, the implementation
is overridden with an instance attribute. When the FileDescriptor
becomes disconnected, that instance attribute is deleted. Thusly,
FileDescriptors which were previously connected but have become
disconnected have a fileno method that raise NotImplementedError,
rather than returning -1.  Much of the time this is not problematic,
because Twisted manages to get those FileDescriptors out of reach
of the reactor before their fileno() is needed again.

This changes the abstract behavior to return -1.  This should result
in absolutely no behavioral changes, aside from the less verbose
error reporting in cases which are not actually errors.