command history for gtk2manhole client
I've refactored some of the guts of the key-event handler to make dispatching
of key events simpler. the addition of a key_(mod)_(key) method will cause
that key or key-combination to be dispatched to that method. 

The history class works properly, albeit simply. there are unittests
that I'm still working on that will be checked in shortly.

In the text-input-widget, you press C-p to get the previous history
item and C-n to get the next. The current input buffer's state is not
saved when you hit C-n or C-p and changes will be lost. duplicate history 
items are allowed for now.

planned for future addition is reverse-history search (akin to bash/zsh's
C-r)