Fix CopiedFailure and CopyableUnslicer
flavors: create a NonCyclicRemoteCopyUnslicer, primarily for use by
CopiedFailure (since it was the Deferred used to handle cyclic RemoteCopy
objects that caused the problems with CopiedFailure). Add a .nonCyclic flag
to the RemoteCopy class (note!: not the instance) which tells us to use it.
Extend registerRemoteCopy to let you register IUnslicers as well as
IRemoteCopy factories.

pb.py: handle IUnslicer and IRemoteCopy as 'copyable' factories. Mark
CopiedFailure as nonCyclic.

test_pb.py: enable RemoteCopy tests now that they work