a fix for the plugin loading problem
it seems that in certain circumstances, sys.path may not contain solely 
string objects. For some weird reason, i ran into the trial bug eveyone's
been complaining about, and when i did sys.path in pdb, there was an empty
list in sys.path! oO

i don't know why this happened, but this is a trivial fix for that situation
so that getPluginFileList doesn't puke.