Fix lore warnings, add whitespace, tweak formatting.
clients.xhtml: skip first 24 lines of ircLogBot.py, to avoid splitting the
big blockquote and keep lore's python syntax-colorer from getting confused.
I'm not sure whether the blockquote was meant to be included or not (I
suspect the skipLines="15" was based upon an earlier version of the example
file), but splitting it is bad.

svn-dev.xhtml: break a long shell command to hush the lore 80-column warning.
Pity, because it's slightly more readable at the 95-ish chars it was at.

newweb.xhtml: replace <h4> with <p> and <strong>, accomplishes something
pretty close.