Merge /b/e/words-update into trunk
Author: exarkun
Reviewer: itamar

This is a complete rewrite of the twisted.words server code.  Botbot and Tendril features are dropped 
for the time being.  Provided is a generic chat server accessible via IRC and PB, and with a plugin 
mechanism for adding new protocols.  Authentication is done via newcred; unlike the previous revision,
there is no support for registering new users through the server.  This should be dealt with using a 
generic checker manipulation system (so that any checker configuration can be manipulated, rather than 
one or two I choose to support in words).

Webwords is also gone, as there are no exciting web technologies in Twisted.  The plugin mechanism is,
however, used to provide a Nevow LivePage interface to the chat service which could be extended to 
perform all the same tasks previously covered by Webwords.

t-im may or may not still work.