Merge threadpool-1385
This cleans up many of the threading tests, fixes an infelicity in 
the treatment of the reactor threadpool by trial, brings Qt reactor
into line with the coding standard, fixes a bug in its install() 
function, and improves test coverage for parts of IReactorThreads.

It also changes reactor threadpool initialization so that threads are 
never started *before* the reactor starts.  This is done by using 
callWhenRunning to start the threadpool, rather than starting it 
synchronously.