Merge waker-1344
This adds a test to make sure that if the reactor thread
ever falls way behind and other threads keep piling requests
on it, callFromThread() will not raise an error when the
socket used to wake the main thread fills up.

This changes the Win32 waker in a similar fashion, but I
am almost positive it does not fix the problem on Windows.
This code is probably no more broken than that which it
replaces, though, and committing to trunk will let us see
what exactly the behavior of the new test is.