Merge cls-type-1581
Author: exarkun
Reviewer: jerub
Fixes #1581

This adds a test for basic cache lookup behavior and fixes a bug in that
code as well.  It also fixes the same bug in the root resolver, but there
are no tests for that code (which is probably okay, because I doubt it
works very well anyway).