Ensure that errors/failures from TestCases are reported in a way that works with pyunit.
Author: spiv

Reviewer: radix

Fixes: #1594

Trial's TestResult class has an interface that's richer than, and incompatible 
with, pyunit's.  This change wraps pyunit TestResults in an adapter that does
a simple translation from the richer interface to pyunit's.  The most obvious 
benefit of this is that test failures and errors in Trial TestCases now get 
noticed when run by pyunit.