Adjust twisted-dev-mode's trial integration to simplify specifying multiple test modules
Author: exarkun
Reviewers: radix, glyph

Previously multiple test-case-name modules were supported by using a hackish -x on
the name of each module after the first and enclosing the entire list in double quotes.
Now, the test-case-name is ignored by emacs and the buffer filename is passed
as the --testmodule argument to trial, which figures out what to do.