Merge defgen-error-handling-1709
Author: exarkun
Reviewer: radix
Fixes #1709

This changes deferredGenerator-wrapped generator functions to only produce Deferreds
which fire with results which are *not* the result of a yield waitForDeferred-wrapped
Deferred.  The common annoying manifestation of this bug is that a generator which
handles an exception from waitForDeferred.getResult() with no subsequent yields
would cause the deferredGenerator's result to be that Failure.