Merge errmsg-1399
When code fails, you often want to indicate the general subsystem where it
failed, and indicate why you are logging an error.  The current idiom is to
call log.msg() followed by log.err(); this is kind of dumb.  This change adds a
second argument to log.err() so that you can report errors more clearly, along
with an error message explaining why the traceback has printed.

Fixes #1399

Author: glyph

Reviewer: PenguinOfDoom