Fix a few data-dropping problems in SUX and Microdom.
Author: radix
Reviewer: exarkun

Fixes #1984

First: SUX was dropping the first byte in a non-conformant XML document on the 
floor, if that first byte wasn't <.
SUX was also improved to 1. care about connectionLost, and end the final state in connectionLost.

MicroDOM's parse function was improved in various ways to drop less data in 
non-conformant documents.