Merge timezone-logging-630
Author: exarkun
Reviewer: jerub
Fixes #630

This replaces usage in the default case of time.strftime by the log module
with a custom time formatting routine.  It also replaces the timezone name
with a numeric offset.

Providing a custom time format string is still supported and will cause the
code to revert to strftime.