Merge fetch-pipeline-1992
Author: exarkun
Reviewer: radix
Fixes #1992

This fixes a state corruption bug in the IMAP4 server which manifest when
a client pipelined three or more FETCH commands which included a request
for message text.  The server is now careful to avoid re-entrancy issues
when handling pipelining.  A test for this case has also been added, and
the IMAP4 server's scheduler is now parameterized.

Thanks to tvachon for pointing out this bug.