Merge epollreactor-1953
Author: therve, exarkun 
Reviewer: PenguinOfDoom, Jerub, jml
Fixes #1953

Add a level-triggered epoll-based reactor.

Also re-order the removal and the shutdown of sockets from the reactor
in the tcp code shared between reactors in order to avoid possibly passing
a no-longer valid file descriptor to a control API such as epoll_ctl() or
kevent().