Use `LoggedSuite`  instead of `startKeepingErrors` to trap logged errors  * Author: jml  * Reviewers: Jerub, exarkun  * Fixes #2173
`startKeepingErrors` has been deprecated for a while, but Trial was still
using it to catch errors that were logged during a test run, but outside 
of any particular test case. 

This branch creates a new kind of `TestSuite` called `LoggedSuite`, which
monitors the log and adds any logged errors to the test result object.