Merge qtreactor-stub-2137
Author: PenguinOfDoom
Reviewer: glyph
Fixes #2137

This adds a reactor plugin which will attempt to import qtreactor and
warn that the name qt is now deprecated.  If qtreactor cannot be
imported it will point to the location from which it can be retrieved.

This also adds a similar backwards-compatibility stub for
twisted.internet.qtreactor for applications which install it themselves
rather than using twistd.