Fix FileDescriptor to not call pauseProducing on non-streaming producers.
This fixes a bug in FileDescriptor's interaction with producers where it would
call pauseProducing() on streaming producers, which are not supposed to be
required to implement such a method.

It also adds a test for the bug in #811, and verifies that streaming producers
will be correctly unpaused in all cases.

Author: zseil, exarkun

Reviewer: glyph

Fixes #2286
Fixes #811