Apply FTPClient pwd/cwd improvement patch
Author: therve
Reviewer: jerub
Fixes #2228

Add two new methods to FTPClient, getDirectory and changeDirectory, which provide a
higher-level interface to the existing pwd and cwd methods and removing the need for
application-level parsing of FTP protocol bytes in order to interpret the results of
those commands.

Also clean up various docstring formatting and epytext markup.