Handle setUpClass / tearDownClass logic outside of `TestCase`
 * Author: jml
 * Reviewers: radix, Jerub
 * Fixes #2303
 * Refs #1870

Trial `TestCase`s with `setUpClass` or `tearDownClass` methods use a different
logic to regular tests. Instead of running each test in a separate instance,
they run all the tests from the same instance.

This branch handles that difference by wrapping `setUpClass`-style tests in
a special suite: `SharedClassSuite`. This suite runs `setUpClass`, 
`tearDownClass` and runs all of its tests from the same instance.

This branch affects `unittest` compatibility in two ways. The first is that 
tests that have `setUpClass` or `tearDownClass` must be run from within 
`SharedClassSuite`. Otherwise, they simply will not work as expected. The 
second is that `_classCleanUp` will only be called if tests are within a 
`ClassSuite` (`SharedClassSuite` is a subclass). This probably doesn't matter 
for most tests. However, it does matter for tests in `test_threads`, for 
example.

`ClassSuite` doesn't particularly care whether its members are all in the 
same class or not. The decision to have an extra-strict cleanup at the end of 
every class is a fairly arbitrary one. `SharedClassSuite` cares deeply about 
the class of its members.