Merge parameterize-looping-call-2633.
Author: radix
Reviewer: therve
Fixes #2633
Fixes #2634

LoopingCall (and code which uses it) is now much more testable because the
scheduler it uses is overridable. In support of this, a 'seconds' method
was added to the reactor (IReactorTime) to strengthen the logical grouping
of 'what is the current scheduling-time' and 'schedule a call for later'.