Add `TestCase.addCleanup` and `defer.runSequentially`.
 * Author: jml
 * Reviewer: therve
 * Fixes #2610

`addCleanup` is an easy way to add cleanup routines to unit tests. It's
particularly useful for factory methods which return objects that require some
sort of cleanup (such as an explicit disconnect). Because this is Twisted,
callables passed to `addCleanup` can return Deferreds.

`runSequentially` takes a sequence of nullary callables and runs them
sequentially. If a callable returns a `Deferred`, `runSequentially` will not
run the next callable until the callback chain of the `Deferred` has been
completed.