Merge amp-factoring-2657+2667.
Author: radix
Reviewer: therve
Fixes #2657
Fixes #2667
Fixes #2656
Fixes #2664

It is now possible to use public methods on Command
objects to serialize and parse arguments and responses
based on that Commandn's arguments and response schemas.
It was already possible to serialize responses via
the Command in this way; analogs for the other three cases
were added. (#2657 & #2667).

Some docstrings in nearby code were alse fixed in this branch.
(#2656 & #2664).