Merge trial-memory-2275
Author: therve
Reviewers: jml, exarkun
Fixes #2275

Add two optmizations to trial to reduce its memory footprint: add a new
DestructiveTestSuite, used by default, that don't keep a reference to test
instances after run, and replace the storage of successfull result by a counter.
This allows the garbage collector to clean a lot of objects, and also speed up
the whole suite.