Merge test-build-protocol-2831
Author: exarkun
Reviewer: therve
Fixes #2831

Rewrite the old test case which covered the value passed to buildProtocol as
the remote address and the values returned by the getPeer and getHost methods
of tcp transports as two new tests which cover the same behavior but do not
have a race condition in their cleanup.  This removes a very intermittent
failure which would leave a DelayedCall and a tcp transport in the reactor
causing trial to report a warning.  The new tests are also more thoroughly
documented and should be easier to maintainer.