Merge epoll-hup-err-2809-2
Author: exarkun
Reviewer: therve
Fixes #2809
Refs #1780
Refs #1662
Refs #2833

Change the error (EPOLLHUP, EPOLLERR) checking code so that it can
handle the case where they are both set, rather than only checking
the cases where one or the other is set.  This allows epollreactor
to properly detect and report connections which are lost uncleanly
and also are not being polled for input events.

The test case for this is slightly convoluted for a number of
reasons:

  * gtkreactor cannot pass the new test, and I am not fixing that
    reactor in this branch.  The test will skip when that reactor
    is in use.  #2833 was created as a result of this in order to
    be able to remove the broken gtkreactor eventually.
  * If a file descriptor is being polled for neither read nor write
    events, it will not be polled for hang-up events either.  This
    is a known issue, #1662.  When it is fixed, the test will need
    to be updated.
  * In order to exercise the case where the event reported from epoll
    is only HUP|ERR (specifically, not IN), the test pauses one of
    the involved transports.  However, due to #1780, it must use a
    timing hack to accomplish this.  Once that issue is resolved,
    it should be possible to remove this hack.