Merge safer-lastwrite-test-2858
Author: exarkun
Reviewer: therve
Fixes #2858

Change the test method covering the reliability of the
very last write on a pty stdout so that it does not fail
if the runtime generates warning messages on stderr which
get mixed up with the test program's output.

Also change stdio_test_lastwrite.py so that it uses the
reactor specified on its command line instead of always
using the default reactor.