Refactor t.c.ssh.keys.* into an object and rewrite tests
Author: z3p
Reviewers: exarkun, glyph, therve
Fixes #2679

twisted.conch.ssh.keys.* was previously a collection of functions that were
poorly documented and poorly tested. This refactors that module's functionality
into a single class: Key. The old functions remain, but are deprecated. The 
tests in test/test_keys have been updated to include the refactoring as well as
increasing the test coverage.