Merge encoder-decoder-signature-2929
Author: exarkun
Reviewer: therve
Fixes #2929

Change the signatures of twisted.mail.imap4.encoder and twistedmail.imap4.decoder
so that they accept a second argument defining the error handling behavior to use.
This makes them compatible with the documentation for codec encoder and decoder
functions and makes the IMAP4-UTF-7 codec registered by the imap4 module work
with the two-argument form of str.encode and str.decode.

The error behavior parameter is currently ignored.