Refactor AMP to separate sending and receiving binary boxes from handling them.
Author: glyph

Reviewer: therve

Fixes #2810

This change does not provide much in the way of new functionality, but it makes a group of previously opaque, private interfaces public and documented.  In so doing, it creates a bunch of new, cleaner tests, and opens up a number of more obvious possibilities for implementing protocol-extension features such as streaming transfers, multiplexing (multiple command-handlers on a single socket), and authentication.