Correct interaction of `namedAny` and import hooks.
The stack-walking code in namedAny was using a broken method (counting fixed numbers of frames) for determining the provenance of an import error.  This would break when Python code was on the stack as a result of an import hook or importer written in Python.

This change fixes the issue with a more reliable way of walking the stack (inspecting frames for telltale variables) and also explicitly separates and documents the purpose of this inspection.

Author: glyph, exarkun

Reviewer: therve, jerub

Fixes #698