Merge pauseproducing-ignored-1780-2
Author: therve
Reviewer: exarkun
Fixes #1780

Allow transport.pauseProducing to work in connectionMade on the client side.
Previously, due to internal mechanisms in tcp, this was incorrectly ignored.

To ease the tests, a new API of the reactor has been added, getReaders and
getWriters, that respectly returned the file descriptors monitored for reading
and writing by the reactor.