Merge system-event-ordering-3198-2
Author: exarkun
Reviewer: therve
Fixes: #3198

Add unit tests which are parameterized on the reactor implementation and
are can be run against any (or all) available reactor, regardless of the
reactor selected at trial invocation time.

Some tests are added for IReactorCore, IReactorTime, IReactorTCP, and
IReactorProcess.  None of these interfaces is completely covered by
the new tests; some old tests have been removed in favor of these new
ones.  Eventually most of the old tests should be converted to this
new style so that all reactors can be tested more easily.

Some small changes have been made to allow new reactors to be created
and used independently of the global reactor.  Since most reactor tests
are still written in the old style, this doesn't mean that multiple
reactors are now supported, only that they work well enough to pass the
small number of new tests which require this.

The "running" attribute is now documented (as most people treated it as
a public attribute already) and more thoroughly tested.