Merge private-epytext-formatting-3268
Author: mwh, exarkun
Reviewer: exarkun, therve
Fixes: #3268

Update the docstrings of a number of private APIs so that they are
valid epytext (this, necessary since we want pydoctor to include
all APIs in the generated documentation so that private APIs which
are re-exposed as public somewhere else get documented).

Also change one of the release unit tests from asserting that private
APIs are not documented to asserting that private APIs are documented.