Merge async-loopback-3302
Author: exarkun
Reviewer: therve
Fixes: #3302

Change the implementation of twisted.protocols.loopback.lookbackAsync
so that it uses `IReactorTime.callLater` to "cut back" the stack when
resuming data delivery in response to a transport write.  This avoids
the case where a protocol's dataReceived method might be called when
it is in the middle of a call to its transport's write method.  Real
transports do not produce this behavior, since they typically must go
back into the event notification loop to learn about the written data.