Merge ssl-disconnection-3218-3
Author: exarkun
Reviewer: jknight, therve
Fixes: #3218

Work around a bad behavior in the Linux implementation of select(2)
which interacted with OpenSSL to cause `SelectReactor` to sometimes
be unable to detect that a TCP connection which had switched to TLS
had been disconnected.  This change causes `SelectReactor` to check
such connections for readability (in addition to the existing check
for writeability), since Linux adds these sockets only to the read-
set.

This also adds the utility method `ReactorBuilder.runReactor` which
wraps `reactor.run()` with additional logic to enforce timeouts for
tests.