Merge starttls-with-buffered-686-3
Author: exarkun
Reviewer: jknight, therve
Fixes: #686

Change the implementation of ITLSTransport.startTLS so that
if there is unsent data in the buffer, TLS negotiation is
delayed until all of that data is sent.  This changes the
existing behavior which would log an error if the data
could not be sent immediately and then drop the connection.