Merge web-wsgi-2753
Author: exarkun
Reviewer: therve
Fixes: #2753

Add a rudamentary WSGI container to `twisted.web`.

`twisted.web.http.Request.setResponseCode` is also changed to be more
strict in the values it accepts, raising a `TypeError` immediately if
given a non-integer for the status code.  Previously this would only
raise an exception later when an attempt was made to log the request.