Merge msnp8-cvr0-3394
Author: exarkun
Reviewer: glyph
Fixes: #3394

Change the MSN protocol implementation so that it does not require
the `CVR0` protocol to be included in the `VER` command.  The value
was never used for anything, although it was always previously
sent by the server.  Now it is no longer sent, so requiring it
only breaks the client.