Merge osvfstest-write-2974
Author: exarkun
Reviewer: therve
Fixes: #2974

Ensure that the test file created in the VFS unit test setup
fixture is flushed and closed before any of the tests try
to use its contents.  This removes the reliance on reference
counting to free the file object and flush the buffer in
order for the tests to pass.