Merge iutils-errors-3239
Author: exarkun
Reviewer: therve
Fixes: #3239

Augment the `IOError` with which `getProcessOutput` can fail when
it receives bytes on stderr but is not expecting to.  The exception
now offers a `processEnded` attribute bound to a `Deferred` which
fires when the offending process ends.  The unit tests now use this
attribute in order to wait long enough so that the process does not
cause trial to report an unclean reactor.