Merge request-body-timeout-1903-2
Author: exarkun
Reviewer: therve
Fixes: #1903

Reset the idle timeout in the HTTP server whenever data from the request
body is received.  Since previously request body data did not cause the
timeout to reset, if a client was making progress on sending a request
body it might still have been disconnected due to the timeout.