Merge header-value-quoting-2346
Author: exarkun, itamar
Reviewer: jknight
Fixes: #2346

Fix generation of HTTP header values in web2 for the case where a parameter value
includes a byte which requires quoting.  Previously these values would not be
quoted, resulting in output which probably could not be parsed.  Now quoting will
be applied if necessary to avoid generating bad output.