Merge amp-key-enforcer-3478
Author: exarkun
Reviewer: glyph
Fixes: #3478

Change `BinaryBoxProtocol` so that as soon as it receives a key length
prefix larger than 255 it closes the connection and reports a protocol
violation as the reason for any failures.  Keys are restricted to 255
or fewer bytes by the protocol; early disconnection for violation of this
simplifies tracking down bugs in protocol implementations.