Merge follow-redirect-clash-3192
Author: jonathanj
Reviewer: exarkun
Fixes: #3192

Fix `getPage` so that the value supplied for `followRedirect` does
not override the value supplied for all previously issued and still
running `getPage` calls.

Previously `HTTPClientFactory` would set the `followRedirect` class
attribute of `HTTPPageGetter` rather than setting it on the one
instance created to satisfy one `getPage` call.  Now it sets the
attribute on an instance instead.