Merge remove-plugins-tml-3246-2
Author: thijs, exarkun, itamarst
Reviewer: radix, thijs
Fixes: #3246

Remove from the documentation all remaining references to the old style of plugins
and replace them with documentation based on the new style of plugins, where
appropriate.  The Lore plugin interface `IProcess` is also slightly improved so
that it at least defines the attributes which are actually required of it and
does not document an attribute which is never used by Lore.