Merge findlinestarts-3598
Author: exarkun
Reviewer: radix, mwhudson
Fixes: #3598

Change the implementation of `twisted.trial.unittest.TestCase.flushWarnings` so
that it works properly when a warning is emitted from a module which is defined
by a source file which has been renamed since it was compiled.  Previously, the
rename would cause file names to disagree in certain places and prevent warnings
emitted from code in a renamed file from being flushed.