Merge xmlrpclib-datetime-hack-2446
Author: exarkun
Reviewer: mwhudson
Fixes: #2446

Add error handling to the fault serialization code in `twisted.web.xmlrpc.XMLRPC` so
that responses are not left hanging indefinitely if there is an error serializing an
error report.  Also insert a monkey patch for `xmlrpclib.DateTime` deserialization so
that these objects will correctly round-trip over XML-RPC on Python 2.3 and Python 2.4.