Merge smtpsender-retry-3638-2
Author: atsuoi, exarkun
Reviewer: therve
Fixes: #3638

Fix the retry support in `SMTPSenderFactory` so that if one or more recipient
addresses is accepted or if part of the message body file is read and sent to
the remote MTA that data is still used on subsequent delivery attempts.  The
previous behavior could exhaust the recipient address iterator on the first
failed attempt or move the position pointer in the message body file causing
subsequent deliver attempts to operate using invalid data.