Make twisted.words IRC server compatible with Pidgin.
author: itamar, tenth

reviewer: therve, exarkun, z3p, glyph

Fixes #2385

While there are other issues to be resolved with MOTD and AWAY behaviors 
as per #3664 and #3668 respectively, twisted.words now always sends an 
MOTD message, which is required for Pidgin to successfully connect to 
it. Also, the "nickname" and "realm" attributes are now initialized in 
time for them to appear in error messages during login.