Merge microdom-ns-3672
Author: shywolf9982, exarkun
Reviewer: exarkun, glyph
Fixes: #3672

Fix microdom's serialization of elements with respect to namespaces.  In
particular, emit the correct closing tag when the corresponding open tag is
qualified with a namespace prefix (previously the namespace prefix would be
omitted from the closing tag) and correctly track which namespace is active
as the default.