Merge imap4-utf7-registration-3663
Author: exarkun
Reviewer: glyph
Fixes: #3663

Update the IMAP4-UTF7 codec registration to use the preferred Python 2.5 behavior
of returning a CodecInfo instance instead of a four tuple.  This causes the
codecs.getreader and codecs.getwriter functions to work for this encoding (again).
Also fix the implementation of the stream writer to define the correct method,
`encode`, instead of the incorrect one, `decode`.