Merge imap4-unsolicited-responses-1105
Author: jhohm, exarkun
Reviewer: glyph
Fixes: #1105

Handle more unsolicited data received from the server instead of
printing it out and otherwise ignoring it.

In particular, now if unsolicited data is received on the same
line as solicited data, it is separated from the solicited data
and delivered to the appropriate callback.  Previously, this case
could cause the entire line to be ignored or it could cause the
unsolicited data to be delivered to the application callback with
the solicited data.