Merge distrib-request-headers-3697
Author: exarkun
Reviewer: therve, glyph
Fixes: #3697

Handle the new Headers instance at twisted.web.http.Request.requestHeaders
specially when serializing a Request over a PB connection (for example,
when sending it to a distrib client) so that the values survive the
transmission rather than morphing into an Unpersistable instance.  This
re-merge of the branch fixes Python 2.3 compatibility.