Simplify testing of XML-RPC failure mode to avoid triggering sgmlop parser bug.
This change changes the data expected to be unparseable to a small, static string, rather than the full contents of the PYC of the test case (which, among other things, contains the bytes of a few valid XMLRPC responses).

This allows Twisted's tests to pass on the upcoming Ubuntu release, Jaunty Jackalope, in its default package configuration.  Note however that removing the `python-reportlab-accel` package will fix the older version of this test as well.

Author: glyph

Reviewer: exarkun

Fixes #3779