Merge conch-client-broken-again-3809
Author: z3p, therve, exarkun
Reviewer: glyph
Fixes: #3809

Fix the mis-use of the SSH agent client by `twisted.conch.client.default`:
the agent's `signData` method expects the key as a string, not an object.
Update the documentation for related APIs which was either missing or which
documented the types of various key-related APIs incorrectly.

This allows the command-line conch client to use an ssh agent successfully
again.