Merge conch-test-skips-3916
Author: exarkun
Reviewer: therve
Fixes: #3916

Change two of Conch's test modules so that they skip when running on Windows
when both pyasn1 and PyCrypto are available.  Previously they would try to import
termios and fail, as this module is not available on Windows.

Also deleted is some left-over support code for Conch connection sharing, a feature
removed some time ago.